+++ cocos2dx/label_nodes/CCLabelAtlas.cpp
@@ -64,7 +64,7 @@ bool CCLabelAtlas::initWithString(const char *string, const char *charMapFile, u
 bool CCLabelAtlas::initWithString(const char *string, CCTexture2D* texture, unsigned int itemWidth, unsigned int itemHeight, unsigned int startCharMap)
 {
     CCAssert(string != NULL, "");
-    if (CCAtlasNode::initWithTexture(texture, itemWidth, itemHeight, strlen(string)))
+    if (CCAtlasNode::initWithTexture(texture, itemWidth, itemHeight, (unsigned int)strlen(string)))
     {
         m_uMapStartChar = startCharMap;
 		this->setText(string);
@@ -115,7 +115,7 @@ bool CCLabelAtlas::initWithString(const char *theString, const char *fntFile)
 //CCLabelAtlas - Atlas generation
 void CCLabelAtlas::updateAtlasValues()
 {
-    unsigned int n = m_sString.length();
+    unsigned int n = (unsigned int)m_sString.length();
 
     ccV3F_C4B_T2F_Quad quad;
 
@@ -186,7 +186,7 @@ void CCLabelAtlas::updateAtlasValues()
 //CCLabelAtlas - CCLabelProtocol
 void CCLabelAtlas::setText(const char *label)
 {
-    unsigned int len = strlen(label);
+    unsigned int len = (unsigned int)strlen(label);
     if (len > m_pTextureAtlas->getTotalQuads())
     {
         m_pTextureAtlas->resizeCapacity(len);
