+++ cocos2dx/support/ccUTF8.cpp
@@ -129,7 +129,7 @@ static const char *const g_utf8_skip = utf8_skip_data;
  * */
 unsigned int cc_utf8_find_last_not_char(std::vector<unsigned short> str, unsigned short c)
 {
-    int len = str.size();
+    int len = (int)str.size();
     
     int i = len - 1;
     for (; i >= 0; --i)
@@ -148,7 +148,7 @@ unsigned int cc_utf8_find_last_not_char(std::vector<unsigned short> str, unsigne
  * */
 static void cc_utf8_trim_from(std::vector<unsigned short>* str, int index)
 {
-    int size = str->size();
+    int size = (int)str->size();
     if (index >= size || index < 0)
         return;
     
@@ -171,7 +171,7 @@ bool isspace_unicode(unsigned short ch)
 
 void cc_utf8_trim_ws(std::vector<unsigned short>* str)
 {
-    int len = str->size();
+    int len = (int)str->size();
     
     if ( len <= 0 )
         return;
@@ -277,7 +277,7 @@ cc_utf8_get_char (const char * p)
 
 unsigned short* cc_utf8_to_utf16(const char* str_old)
 {
-    int len = cc_utf8_strlen(str_old, -1);
+    int len = (int)cc_utf8_strlen(str_old, -1);
     
     unsigned short* str_new = new unsigned short[len + 1];
     str_new[len] = 0;
