+++ extensions/network/HttpRequest.h
@@ -121,7 +121,7 @@ public:
     /** Get the size of request data back */
     inline int getRequestDataSize()
     {
-        return _requestData.size();
+        return (int)_requestData.size();
     }
     
     /** Option field. You can set a string tag to identify your request, this tag can be found in HttpResponse->getHttpRequest->getTag()
